refactor TODOMenu to have states inside the componment, refactor folders
previously the states used to create the todo were stored at top level, as the
refactoring of the menu followed later and it was misunderstood the necessity to
keep the states at top level. this has been changed and now the states such as
name/setname, priority/setpriority, id/setId, date/setDate are stored in TODOMenu
component. the size of the TODOPropsPackage has been changed to only pass around the
todo/setTodo

the folders with the modules have been changed to better show the hierarchy of the
components in the TODO.tsx: all the components are either stored in
TODOMenuComponents or TODOListComponents