add filter function
the todoList of iTodo[] is moved to is own component and get sorted by day, month and year on a chain of .sort()
before being passed to .map() and become of type JSX.Element[] | undefined. the map
function get a string from a component where i can select to filter for today, this
week or remove the filter. a switch case will determine which element will be part of the todoList array depending on the date