fix TODOListItemsFiltered
previously the array was sorted, then mapped (as: array.sort(sort by date).map(return by today or week or clear )) but map has to return an element no matter what, so this would result in a full array filled with undefined everitime a filter was set with data that satisfy an empty set (eg: nothing with the date of today).this has been fixed by applying a .filter() at the end that only makes sure that the final data filtered doesnt have any undefined data ( eg: .filter(item=> item != undefined)). this way it was possible to set up a ternary operator in the return of the component that will return the string on screen "No items to show for this filter selection" if the lenght of the returned array after the .sort().map().filter() === 0. This is done to ease up the tesing, as i shoudlnt test if the lenght of the returned item on screen is 0, but if that kind of string is present

TODOFilter.test.jsx added a new test to check when the filtering return no element on screen. an element with the date of +8 from today is added, then the filter is set for Today and Week, in both case the string for empty array is returned. when set for Clear Filter, one item is present