BUGFIX: fix expired date
old system would just compare if the date (just by the day) with a new date object (just the day) and if this is less than the previous, resulting that an amonut far in time might be less thatn the previous date, resulting in an expired data warning. this has been resolved by removing getDate() from both comparison, now it compares the full data.

there might be an issue where if i set the data of a couple of years from now i might get it back tomorrow with the current year: i checked the database and the year is set correctly, so i will need to check again